requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0
ollama>=0.1.7
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0 